requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
faiss-cpu
